lookup: Support tbid option
Support bpf_fib_lookup with tbid option. We expose table option similar
to the iproute2. It requires an additional flag BFP_FIB_LOOKUP_TBID to
distinguish table 0 and unspecified, but we don't expose the flag
explicitly. Since we can distinguish them with the table option
specified vs unspecifid. When the table option is specified, we set it
transparently with BPF_FIB_LOOKUP_DIRECT.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>